+++ src/test/java/com/sefford/brender/data/DefaultAdapterDataTest.java
@@ -24,20 +24,26 @@ import org.mockito.Mock;
 import org.robolectric.RobolectricTestRunner;
 
 import java.util.ArrayList;
+import java.util.List;
 
 import static org.hamcrest.CoreMatchers.instanceOf;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThat;
+import static org.mockito.Mockito.when;
 import static org.mockito.MockitoAnnotations.initMocks;
 
 @RunWith(RobolectricTestRunner.class)
 public class DefaultAdapterDataTest {
 
-    private static final int EXPECTED_POS = 2;
+    static final int EXPECTED_POS = 2;
+    static final int EXPECTED_VIEWTYPE_COUNT = 3;
+    static final int EXPECTED_RENDERABLE_ID = 991984;
+
     @Mock
-    Renderable element;
+    Renderable renderable;
 
     ArrayList<Renderable> master;
+
     DefaultAdapterData adapterData;
 
     @Before
@@ -45,9 +51,9 @@ public class DefaultAdapterDataTest {
         initMocks(this);
 
         master = new ArrayList<>();
-        master.add(element);
-        master.add(element);
-        master.add(element);
+        master.add(renderable);
+        master.add(renderable);
+        master.add(renderable);
 
         adapterData = new DefaultAdapterData(master);
     }
@@ -64,11 +70,58 @@ public class DefaultAdapterDataTest {
 
     @Test
     public void testGetItem() throws Exception {
-        assertEquals(element, adapterData.getItem(EXPECTED_POS));
+        assertEquals(renderable, adapterData.getItem(EXPECTED_POS));
     }
 
     @Test
     public void testGetFilter() throws Exception {
         assertThat(adapterData.getFilter(), instanceOf(NullFilter.class));
+    }
+
+
+    @Test
+    public void testComputeViewTypes() throws Exception {
+        when(renderable.getRenderableId()).thenReturn(EXPECTED_RENDERABLE_ID)
+                .thenReturn(EXPECTED_RENDERABLE_ID + 1000)
+                .thenReturn(EXPECTED_RENDERABLE_ID + 2000);
+        List<Renderable> renderables = new ArrayList<>();
+        renderables.add(renderable);
+        renderables.add(renderable);
+        renderables.add(renderable);
+        adapterData.viewTypes.clear();
+
+        adapterData.computeViewTypes(renderables);
+
+        assertEquals(EXPECTED_VIEWTYPE_COUNT, adapterData.viewTypes.size());
+    }
+
+    @Test
+    public void testGetViewTypeCount() throws Exception {
+        when(renderable.getRenderableId()).thenReturn(EXPECTED_RENDERABLE_ID)
+                .thenReturn(EXPECTED_RENDERABLE_ID + 1000)
+                .thenReturn(EXPECTED_RENDERABLE_ID + 2000);
+        final List<Renderable> renderables = new ArrayList<>();
+        renderables.add(renderable);
+        renderables.add(renderable);
+        renderables.add(renderable);
+
+        adapterData = new DefaultAdapterData(renderables);
+
+        assertEquals(EXPECTED_VIEWTYPE_COUNT, adapterData.getViewTypeCount());
+    }
+
+    @Test
+    public void testNotifyDataSetChanged() throws Exception {
+        master.clear();
+        when(renderable.getRenderableId()).thenReturn(EXPECTED_RENDERABLE_ID)
+                .thenReturn(EXPECTED_RENDERABLE_ID + 1000)
+                .thenReturn(EXPECTED_RENDERABLE_ID + 2000);
+        master.add(renderable);
+        master.add(renderable);
+        master.add(renderable);
+
+        adapterData.notifyDataSetChanged();
+
+        assertEquals(EXPECTED_VIEWTYPE_COUNT, adapterData.getViewTypeCount());
     }
 }